horizontalOffset in settings
And add tolerance to Field so that -horizontalOffset can be treated as
tolerance and in this way the drawing still reaches the right side even
though it is shifted to the left.

Actually it would probably be a better idea to simply make Field size a
litle bigger instead of using tolerances. I may still do that.